♻️ refactor: move cond_reverse to galax.potential
`cond_reverse` is a generic jax helper with one consumer in `potential` and one
in `dynamics`, but it lived in `galax.dynamics._src.utils`, so
`potential/_src/xfm/translate.py` imported it *upward* at module level -- the
only module-level `potential` -> `dynamics` import in the codebase.

Moving it to `galax.potential._src.utils` reverses that: `dynamics` now imports
it downward, which it already does for `coord_dispatcher` and `speed_of_light`
from the same module.

`T_co` goes with it: it was defined next to these helpers and never used.

Split out of #799: it stands on its own and doesn't depend on the namespace cut.
It is load-bearing there -- once the type aliases are split per subpackage, this
import becomes a hard `ImportError` on a partially initialised module -- but the
decoupling is worth having regardless.
