+++ src/galax/potential/_src/xfm/translate.py
@@ -19,10 +19,10 @@ from unxt.quantity import AllowValue
 
 import galax._custom_types as gt
 from .base import AbstractTransformedPotential
-from galax.dynamics._src.utils import cond_reverse
 from galax.potential._src.base import AbstractPotential
 from galax.potential._src.params.base import AbstractParameter
 from galax.potential._src.params.field import ParameterField
+from galax.potential._src.utils import cond_reverse
 
 
 @final
